modify relay structure
Former-commit-id: 24a265a0dee0f5b411b23c2ac41aa769e9043821
Former-commit-id: 497a2810fa088a80b35011e81ed76637583eab99

+++ cmd/relayd/relayd/relayd.go
@@ -200,7 +200,7 @@ func (r *Relayd) dealOrder() {
 
 func (r *Relayd) requestRelayOrders(status types.RelayOrderStatus) (*types.QueryRelayOrderResult, error) {
 	payLoad := types.Encode(&types.QueryRelayOrderParam{
-		Status: status,
+		Status: int32(status),
 	})
 	query := types.Query{
 		Execer:   executor,
